レビュー対応:共通化 #82
