+++ src/features/home/HomeScreen.tsx
@@ -1,9 +1,7 @@
-import { apiClient } from '@/src/api';
-import type { MemoListItemResponse } from '@/src/api/generated/apiSchema';
 import { ConfirmDialog } from '@/src/shared/components';
+import { useDeleteMemoFlow } from '@/src/shared/hooks/useDeleteMemoFlow';
 import { router } from 'expo-router';
-import { useCallback, useState } from 'react';
-import { Alert, ScrollView, View } from 'react-native';
+import { ScrollView, View } from 'react-native';
 
 import { FolderList } from './folder-list';
 import { RecentNotes, useRecentMemos } from './recent-notes';
@@ -11,9 +9,13 @@ import { RecordFab } from './record-fab';
 
 export function HomeScreen() {
   const { memos, isLoading, error, refresh } = useRecentMemos();
-  const [memoToDelete, setMemoToDelete] = useState<MemoListItemResponse | null>(null);
-  const [isDeleteDialogVisible, setIsDeleteDialogVisible] = useState(false);
-  const [isDeleting, setIsDeleting] = useState(false);
+  const {
+    memoToDelete,
+    isDeleteDialogVisible,
+    handleDeleteRequest,
+    handleDeleteCancel,
+    handleDeleteConfirm,
+  } = useDeleteMemoFlow({ onDeleted: refresh });
 
   const handleMemoClick = (memoId: string) => {
     router.push(`/note/${memoId}`);
@@ -22,34 +24,6 @@ export function HomeScreen() {
   const handleStartRecording = () => {
     router.push('/record');
   };
-
-  const handleDeleteRequest = useCallback((memo: MemoListItemResponse) => {
-    setMemoToDelete(memo);
-    setIsDeleteDialogVisible(true);
-  }, []);
-
-  const handleDeleteCancel = useCallback(() => {
-    if (isDeleting) return;
-    setIsDeleteDialogVisible(false);
-    setMemoToDelete(null);
-  }, [isDeleting]);
-
-  const handleDeleteConfirm = useCallback(async () => {
-    if (!memoToDelete || isDeleting) return;
-
-    setIsDeleting(true);
-    try {
-      await apiClient.deleteMemo(memoToDelete.memoId);
-      setIsDeleteDialogVisible(false);
-      setMemoToDelete(null);
-      await refresh();
-    } catch (err) {
-      if (__DEV__) console.error('Failed to delete memo:', err);
-      Alert.alert('エラー', 'メモの削除に失敗しました。もう一度お試しください。');
-    } finally {
-      setIsDeleting(false);
-    }
-  }, [memoToDelete, isDeleting, refresh]);
 
   return (
     <View className="flex-1 bg-t-bg-secondary">
